// apps/ledger-live-desktop/src/newArch/features/ModularDrawer/screens/AssetSelection/modules/types.ts
import type { AssetType } from "@ledgerhq/react-ui/pre-ldls/index";

export type ProviderBalanceAsset = {
  id: string;
  name: string;
  ticker: string;
  balance: string;
  fiatValue: string;
  sortValue: number;
};

export type ProviderBalanceResultsMap = Map<string, ProviderBalanceAsset>;

export type AssetWithBalance = AssetType & {
  sortValue?: number;
};